make metadata wrapper to add extra default mate data
