+++ src/sm_bluesky/common/helper/__init__.py
@@ -1,0 +1,3 @@
+from .add_meta import add_default_metadata
+
+__all__ = ["add_default_metadata"]
