+++ src/sm_bluesky/common/helper/add_meta.py
@@ -4,22 +4,21 @@ from typing import TypeVar, cast
 
 from blueapi.core import MsgGenerator
 
-TCallable = TypeVar("TCallable", bound=Callable)
+TCallable = TypeVar("TCallable", bound=Callable[..., MsgGenerator])
 
 
-P99_DEFAULT_METADATA = {"energy": 1.8, "detector_dist": 88}
-
-
-def add_default_metadata(funcs: TCallable) -> TCallable:
+def add_default_metadata(
+    funcs: TCallable, extra_metadata: dict | None = None
+) -> TCallable:
     @wraps(funcs)
     def inner(
         *args,
         **kwargs,
     ) -> MsgGenerator:
         if "md" in kwargs:
-            kwargs["md"].update(P99_DEFAULT_METADATA)
+            kwargs["md"].update(extra_metadata)
         else:
-            kwargs["md"] = P99_DEFAULT_METADATA
+            kwargs["md"] = extra_metadata
         yield from (funcs(*args, **kwargs))
 
     return cast(TCallable, inner)
